requirements.txt
torchmetrics
transforms
numpy
torchmetrics
pillow
tensorboard
scikit-image 
tqdm 
matplotlib